Fix various lint issues
